+++ config/sitemap.rb
@@ -21,8 +21,11 @@ SitemapGenerator::Sitemap.create do
   add about_path, changefreq: "monthly", priority: 0.9
 
   Developer.visible.newest_first.find_each do |developer|
-    id = Feature.enabled?(:obfuscate_developer_urls) ? developer.hashid : developer.id
-    add developer_path(id:), changefreq: "always", priority: 0.8, lastmod: developer.updated_at
+    add developer_path(id: developer.hashid), changefreq: "always", priority: 0.8, lastmod: developer.updated_at
+
+    if Feature.enabled?(:redirect_db_id_profiles)
+      add developer_path(id: developer.id), changefreq: "always", priority: 0.8, lastmod: developer.updated_at
+    end
   end
 
   add new_user_session_path, changefreq: "weekly", priority: 0.7